move search-ui view components into own file
