+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/engine/engine_search_preview/search_ui_components.tsx
@@ -1,0 +1,99 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+
+import {
+  EuiFieldSearch,
+  EuiBadge,
+  EuiBasicTable,
+  EuiBasicTableColumn,
+  EuiButton,
+  EuiFlexGroup,
+  EuiFlexItem,
+  EuiPanel,
+} from '@elastic/eui';
+import type {
+  InputViewProps,
+  ResultViewProps,
+  ResultsViewProps,
+} from '@elastic/react-search-ui-views';
+
+export const ResultsView: React.FC<ResultsViewProps> = ({ children }) => {
+  return <EuiFlexGroup direction="column">{children}</EuiFlexGroup>;
+};
+
+export const ResultView: React.FC<ResultViewProps> = ({ result }) => {
+  const fields = Object.entries(result)
+    .filter(([key]) => !key.startsWith('_') && key !== 'id')
+    .map(([key, value]) => {
+      return {
+        name: key,
+        value: value.raw,
+      };
+    });
+
+  const {
+    _meta: {
+      id: encodedId,
+      rawHit: { _index: index },
+    },
+  } = result;
+
+  const [, id] = JSON.parse(atob(encodedId));
+
+  const columns: Array<EuiBasicTableColumn<SearchResult>> = [
+    {
+      field: 'name',
+      name: 'name',
+      render: (name: string) => {
+        return <code>&quot;{name}&quot;</code>;
+      },
+      truncateText: true,
+      width: '20%',
+    },
+    {
+      field: 'value',
+      name: 'value',
+      render: (value: string) => value,
+    },
+  ];
+
+  return (
+    <EuiPanel paddingSize="m">
+      <EuiFlexGroup direction="column">
+        <EuiFlexGroup justifyContent="spaceBetween">
+          <code>ID: {id}</code>
+          <EuiFlexItem grow={false}>
+            <EuiFlexGroup gutterSize="xs" alignItems="center">
+              <code>from</code>
+              <EuiBadge color="hollow">{index}</EuiBadge>
+            </EuiFlexGroup>
+          </EuiFlexItem>
+        </EuiFlexGroup>
+        <EuiBasicTable items={fields} columns={columns} />
+      </EuiFlexGroup>
+    </EuiPanel>
+  );
+};
+
+export const InputView: React.FC<InputViewProps> = ({ getInputProps }) => {
+  return (
+    <EuiFlexGroup gutterSize="s">
+      <EuiFieldSearch
+        fullWidth
+        placeholder="search"
+        {...getInputProps({})}
+        isClearable
+        aria-label="Search Input"
+      />
+      <EuiButton type="submit" color="primary" fill>
+        Search
+      </EuiButton>
+    </EuiFlexGroup>
+  );
+};
